refactor: make CoreHeader a jsx file
